first-agent-project/requirements.txt
streamlit
pypdf
langgraph
langchain-google-genai
langchain-community
pinecone
python-dotenv
